Londrina:16.08.2021_02:23hs

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    //Uma outra forma de lidar com a manipulação dos models dentro dos controllers
+    //Construtor injetando uma instância do model(um objeto) nesse controller
+    //"Type Hinting"(com parâmetro "tipado")
     public function __construct(Marca $marca)
     {
         $this->marca = $marca;
@@ -18,9 +21,13 @@ class MarcaController extends Controller
      */
     public function index()
     {
+        /* Assim usando o método estático all() */
         /* $marcas = Marca::all(); */
+        /* Agora acessando o método de "um objeto" */
         $marca = $this->marca->all();
-        return $marca;
+        /* Usando o helper "response()", para modificar os detalhes da resposta do
+            status code http, que será dada pelo laravel. Como 2º parâmetro, o código http */
+        return response()->json($marca, 200);
     }
 
     /**
@@ -33,12 +40,16 @@ class MarcaController extends Controller
     {
         /* De um modo "massivo" usando o "Model": */
         /* Na verdade, "$request->all()" é(retorna)um array associativo */
+        /* Assim usando o método estático create() */
         /* $marca = Marca::create($request->all()); */
+        /* Agora acessando o método de "um objeto" */
         $marca = $this->marca->create($request->all());
         /* dd($marca); */
         /* dd($request->all()); */
         /* return 'Chegamos até aqui (Store)'; */
-        return $marca;
+        /* Usando o helper "response()", para modificar os detalhes da resposta do
+           status code http, que será dada pelo laravel. Como 2º parâmetro, o código http */
+        return response()->json($marca, 201);
     }
 
     /**
@@ -50,40 +61,62 @@ class MarcaController extends Controller
     /* public function show(Marca $marca) */
     public function show($id)
     {
+        /* Acessando o método de "um objeto" */
         $marca = $this->marca->find($id);
-        return $marca;
+        /* Validando: */
+        if ($marca === null) {  /* operador idêntico "===": mesmo tipo e valor */
+            /* return ['êrro' => 'O recurso pesquisado não existe!']; */
+            /* Usando o helper "response()", para modificar os detalhes da resposta do
+               status code http, que será dada pelo laravel. Como 2º parâmetro, o código http */
+            return response()->json(['êrro' => 'O recurso pesquisado não existe!'], 404);
+        }
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    /* public function update(Request $request, Marca $marca) */
+    public function update(Request $request, $id)
     {
-       /* return 'Chegamos até aqui (Update)'; */
-       /* print_r($request->all()); */ //Os dados atualizados do "body" da requisição, deste id.
-      /*  echo '<hr>'; */
-       /* print_r($marca->getAttributes()); */ //Os dados antigos do objeto instanciado, deste id.
-       $marca->update($request->all());
-       return $marca;
+        /* print_r($request->all()); */ //Os dados "atualizados" do "body" da requisição, deste id.
+        /* print_r($marca->getAttributes()); */ //Os dados "antigos" do objeto instanciado, deste id.
+        /* Acessando o método de "um objeto" */
+        $marca = $this->marca->find($id);
+        /* Validando: */
+        if ($marca === null) {  /* operador idêntico "===": mesmo tipo e valor */
+            /* Usando o helper "response()", para modificar os detalhes da resposta do
+               status code http, que será dada pelo laravel. Como 2º parâmetro, o código http */
+            return response()->json(['êrro' => 'O recurso a ser atualizado não existe!'], 404);
+        }
+        $marca->update($request->all());
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
         /* print_r($marca->getAttributes()); */
-        /* return 'Chegamos até aqui (delete)'; */
-        // Recuperando o objeto(id), cuja instância(Model Marca) veio no parâmetro
-        // e executando o método "delete()":
+        // Recuperando o objeto(seu id), cuja instância(Model Marca) veio no parâmetro
+        /* Acessando o método de "um objeto" */
+        $marca = $this->marca->find($id);
+        /* Validando: */
+        if ($marca === null) {  /* operador idêntico "===": mesmo tipo e valor */
+            /* Usando o helper "response()", para modificar os detalhes da resposta do
+               status code http, que será dada pelo laravel. Como 2º parâmetro, o código http */
+            return response()->json(['êrro' => 'O recurso a ser excluido não existe!'], 404);
+        }
+        /* Executando o método "delete()": */
         $marca->delete();
-        return ['msg'=>'A marca foi removida com sucesso']; //Retornando um array associativo.
+        return response()->json(['msg' => 'A marca foi removida com sucesso'], 200) ; //Retornando um array associativo.
     }
 }
